Split ElementType.cs to Duplicate.cs and Similar.cs
